+++ header.php
@@ -6,7 +6,7 @@
  */ ?>
 
 <!DOCTYPE html>
-<html class="<?php echo esc_attr( oceanwp_html_classes() ); ?>" <?php language_attributes(); ?><?php oceanwp_schema_markup( 'html' ); ?>>
+<html class="<?php echo esc_attr( oceanwp_html_classes() ); ?>" <?php language_attributes(); ?>>
 <head>
 	<meta charset="<?php bloginfo( 'charset' ); ?>">
 	<link rel="profile" href="https://gmpg.org/xfn/11">
@@ -16,9 +16,13 @@
 
 <body <?php body_class(); ?> <?php oceanwp_schema_markup( 'html' ); ?>>
 
+	<?php wp_body_open(); ?>
+
 	<?php do_action( 'ocean_before_outer_wrap' ); ?>
 
 	<div id="outer-wrap" class="site clr">
+
+		<a class="skip-link screen-reader-text" href="#main"><?php esc_html_e( 'Skip to content', 'oceanwp' ); ?></a>
 
 		<?php do_action( 'ocean_before_wrap' ); ?>
 
@@ -29,7 +33,7 @@
 			<?php do_action( 'ocean_header' ); ?>
 
 			<?php do_action( 'ocean_before_main' ); ?>
-			
-			<main id="main" class="site-main clr"<?php oceanwp_schema_markup( 'main' ); ?>>
+
+			<main id="main" class="site-main clr"<?php oceanwp_schema_markup( 'main' ); ?> role="main">
 
 				<?php do_action( 'ocean_page_header' ); ?>