+++ partials/footer/layout.php
@@ -10,7 +10,7 @@ if ( ! defined( 'ABSPATH' ) ) {
     exit;
 } ?>
 
-<footer id="footer" class="<?php echo esc_attr( oceanwp_footer_classes() ); ?>"<?php oceanwp_schema_markup( 'footer' ); ?>>
+<footer id="footer" class="<?php echo esc_attr( oceanwp_footer_classes() ); ?>"<?php oceanwp_schema_markup( 'footer' ); ?> role="contentinfo">
 
     <?php do_action( 'ocean_before_footer_inner' ); ?>
 
@@ -26,7 +26,7 @@ if ( ! defined( 'ABSPATH' ) ) {
         if ( oceanwp_display_footer_bottom() ) {
         	get_template_part( 'partials/footer/copyright' );
         } ?>
-        
+
     </div><!-- #footer-inner -->
 
     <?php do_action( 'ocean_after_footer_inner' ); ?>