+++ partials/header/logo.php
@@ -13,7 +13,14 @@ if ( ! defined( 'ABSPATH' ) ) {
 // Vars
 $retina_logo 		= oceanwp_header_retina_logo_setting();
 $full_screen_logo 	= get_theme_mod( 'ocean_full_screen_header_logo' );
-$responsive_logo 	= get_theme_mod( 'ocean_responsive_logo' ); ?>
+$responsive_logo 	= get_theme_mod( 'ocean_responsive_logo' );
+$header_text_color	= null;
+
+if ( display_header_text() && ! OCEAN_EXTRA_ACTIVE && ! class_exists( 'Ocean_Extra_Theme_Panel' ) ) {
+	$header_text_color = ' style="color:#' . get_header_textcolor() . ';"';
+}
+
+?>
 
 <?php do_action( 'ocean_before_logo' ); ?>
 
@@ -56,9 +63,12 @@ $responsive_logo 	= get_theme_mod( 'ocean_responsive_logo' ); ?>
 
 		} else {
 			if (display_header_text()==true){ ?>
-				<a href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home" class="site-title site-logo-text"><?php echo esc_html( get_bloginfo( 'name' ) ); ?></a>
-			<?php } ?>
-		<?php } ?>
+				<a href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home" class="site-title site-logo-text" <?php echo $header_text_color; ?>><?php echo esc_html( get_bloginfo( 'name' ) ); ?></a>
+				<?php
+				do_action( 'ocean_after_site_title' );
+			 }
+
+		} ?>
 
 	</div><!-- #site-logo-inner -->
 
@@ -69,7 +79,7 @@ $responsive_logo 	= get_theme_mod( 'ocean_responsive_logo' ); ?>
 	if (display_header_text()==true){
 		if ( 'top' == oceanwp_header_style()
 			&& '' != get_bloginfo( 'description' ) ) { ?>
-			<div id="site-description"><h2><?php echo bloginfo( 'description' ); ?></h2></div>
+			<div id="site-description"><h2 <?php echo $header_text_color; ?>><?php echo bloginfo( 'description' ); ?></h2></div>
 	<?php }} ?>
 
 </div><!-- #site-logo -->