+++ partials/mobile/mobile-search.php
@@ -14,7 +14,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 $post_type = get_theme_mod( 'ocean_menu_search_source', 'any' ); ?>
 
 <div id="mobile-menu-search" class="clr">
-	<form method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="mobile-searchform">
+	<form method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="mobile-searchform" role="search" aria-label="<?php _e( 'Mobile Search', 'oceanwp' ); ?>">
 		<input type="search" name="s" autocomplete="off" placeholder="<?php echo esc_attr( apply_filters( 'ocean_mobile_searchform_placeholder', __( 'Search', 'oceanwp' ) ) ); ?>" />
 		<button type="submit" class="searchform-submit">
 			<i class="icon icon-magnifier"></i>